+++ components/sync15/src/client/sync_multiple.rs
@@ -8,7 +8,7 @@
 use super::state::{EngineChangesNeeded, GlobalState, PersistedGlobalState, SetupStateMachine};
 use super::status::{ServiceStatus, SyncResult};
 use super::storage_client::{BackoffListener, Sync15StorageClient, Sync15StorageClientInit};
-use crate::clients::{self, CommandProcessor, CLIENTS_TTL_REFRESH};
+use crate::clients_engine::{self, CommandProcessor, CLIENTS_TTL_REFRESH};
 use crate::engine::{EngineSyncAssociation, SyncEngine};
 use crate::error::Error;
 use crate::telemetry;
@@ -229,7 +229,7 @@ impl<'info, 'res, 'pgs, 'mcs> SyncMultipleDriver<'info, 'res, 'pgs, 'mcs> {
         let clients_engine = if let Some(command_processor) = self.command_processor {
             log::info!("Synchronizing clients engine");
             let should_refresh = self.mem_cached_state.should_refresh_client();
-            let mut engine = clients::Engine::new(command_processor, self.interruptee);
+            let mut engine = clients_engine::Engine::new(command_processor, self.interruptee);
             if let Err(e) = engine.sync(
                 &client_info.client,
                 &global_state,
@@ -290,7 +290,7 @@ impl<'info, 'res, 'pgs, 'mcs> SyncMultipleDriver<'info, 'res, 'pgs, 'mcs> {
         &mut self,
         client_info: &ClientInfo,
         global_state: &mut GlobalState,
-        clients: Option<&clients::Engine<'_>>,
+        clients: Option<&clients_engine::Engine<'_>>,
     ) -> telemetry::SyncTelemetry {
         let mut telem_sync = telemetry::SyncTelemetry::new();
         for engine in self.engines {
